backend/requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
PyPDF2
python-dotenv
langchain-google-genai
google-generativeai
langchain-community
